Handle $PAGER being a command with args
If `$PAGER` is set to a command with arguments (e.g. `less -Rs`),
default-pager tried to execute the whole call as a binary. With this
change, the `$PAGER` call will be split into command and arguments.

It's still rather naive, won't handle arguments with escaped spaces or
in quotes but it works for single-character arguments such as those used
by `less`.